SNOW-1063347: Remove series.py (#2205)
<!---
Please answer these questions before creating your pull request. Thanks!
--->

1. Which Jira issue is this PR addressing? Make sure that there is an
accompanying issue to your PR.

   <!---
   In this section, please add a Snowflake Jira issue number.
   
Note that if a corresponding GitHub issue exists, you should still
include
   the Snowflake Jira issue number. For example, for GitHub issue
https://github.com/snowflakedb/snowpark-python/issues/1400, you should
   add "SNOW-1335071" here.
    --->

   Fixes SNOW-1063347

2. Fill out the following pre-review checklist:

- [ ] I am adding a new automated test(s) to verify correctness of my
new code
- [ ] If this test skips Local Testing mode, I'm requesting review from
@snowflakedb/local-testing
   - [ ] I am adding new logging messages
   - [ ] I am adding a new telemetry message
   - [ ] I am adding new credentials
   - [ ] I am adding a new dependency
- [ ] If this is a new feature/behavior, I'm adding the Local Testing
parity changes.

3. Please describe how your code solves the related issue.

This PR removes series.py (following up #2167). As before, any preserved
overrides are given a reason in code comments. The following methods
have been added to `series_overrides.py`:
- __init__
- sub/subtract, and most other binary operators
- aggregate/agg
- apply
- map
- argmax/argmin
- case_when
- dropna
- duplicated
- sum/std/var
- cat/str/dt
- name
- empty
- fillna
- groupby
- info
- _qcut
- replace
- reset_index
- set_axis
- rename
- sort_values
- unstack
- where
- value_counts
- shift
- squeeze
- _prepare_inter_op
- _to_datetime
- to_list
- _create_or_update_from_compiler
- to_frame
- _to_pandas
- to_numpy
- __setitem__
- __repr__

Because of differences in `__getattr__` logic, we need to add `dt`,
`str`, and `columns` to `_ATTRS_NO_LOOKUP` in this PR to prevent
additional queries from being incurred. This will create slightly
different behavior for attempting to access a named index element
directly:
```
>>> s = pd.Series(["a", "b"], index=["str", "columns"])
>>> s.columns
AttributeError: 'Series' object has no attribute 'columns'
>>> s.to_pandas().columns  # DIFFERENT: treats as index access
'b'
>>> s.str
<modin.pandas.series_utils.StringMethods object at 0x135f57940>
>>> s.to_pandas().str  # SAME: gets accessor
<pandas.core.strings.accessor.StringMethods object at 0x135f93b50>
```
