--- src/snowflake/snowpark/modin/pandas/series_utils.py
@@ -1,80 +1,0 @@
-#
-# Copyright (c) 2012-2024 Snowflake Computing Inc. All rights reserved.
-#
-
-# Licensed to Modin Development Team under one or more contributor license agreements.
-# See the NOTICE file distributed with this work for additional information regarding
-# copyright ownership.  The Modin Development Team licenses this file to you under the
-# Apache License, Version 2.0 (the "License"); you may not use this file except in
-# compliance with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific language
-# governing permissions and limitations under the License.
-
-# Code in this file may constitute partial or total reimplementation, or modification of
-# existing code originally distributed by the Modin project, under the Apache License,
-# Version 2.0.
-
-"""
-Overrides the `Series.cat` accessor. `Series.str` and `Series.dt` are inherited from modin.
-"""
-
-# add this line to enable doc tests to run
-from snowflake.snowpark.modin import pandas as pd  # noqa: F401
-from snowflake.snowpark.modin.plugin.utils.error_message import ErrorMessage
-
-
-class CategoryMethods:
-    # CategoricalDType is not supported with Snowpark pandas API. Mark all methods
-    # to be unsupported.
-    category_not_supported_message = "CategoricalDType and corresponding methods is not available in Snowpark pandas API yet!"
-
-    def __init__(self, series) -> None:
-        self._series = series
-        self._query_compiler = series._query_compiler
-
-    @property
-    def categories(self):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    @categories.setter
-    def categories(self, categories):
-        ErrorMessage.not_implemented(
-            self.category_not_supported_message
-        )  # pragma: no cover
-
-    @property
-    def ordered(self):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    @property
-    def codes(self):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def rename_categories(self, new_categories, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def reorder_categories(self, new_categories, ordered=None, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def add_categories(self, new_categories, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def remove_categories(self, removals, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def remove_unused_categories(self, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def set_categories(self, new_categories, ordered=None, rename=False, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def as_ordered(self, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
-
-    def as_unordered(self, inplace=False):
-        ErrorMessage.not_implemented(self.category_not_supported_message)
