+++ src/snowflake/snowpark/modin/pandas/utils.py
@@ -533,7 +533,8 @@ def extract_validate_and_try_convert_named_aggs_from_kwargs(
         A dictionary mapping columns to a tuple containing the aggregation to perform, as well
         as the pandas label to give the aggregated column.
     """
-    from snowflake.snowpark.modin.pandas import Series
+    from modin.pandas import Series
+
     from snowflake.snowpark.modin.pandas.groupby import SeriesGroupBy
 
     is_series_like = isinstance(obj, (Series, SeriesGroupBy))
