+++ src/snowflake/snowpark/modin/plugin/_internal/groupby_utils.py
@@ -36,7 +36,7 @@ from snowflake.snowpark.modin.utils import hashable
 from snowflake.snowpark.window import Window
 
 BaseInternalKeyType = Union[
-    "snowflake.snowpark.modin.pandas.series.Series",  # type: ignore [name-defined] # noqa: F821
+    "modin.pandas.Series",  # type: ignore [name-defined] # noqa: F821
     Hashable,
 ]
 
@@ -66,7 +66,7 @@ def is_groupby_value_label_like(val: Any) -> bool:
     """
     Check if the groupby value can be treated as pandas label.
     """
-    from snowflake.snowpark.modin.pandas.series import Series
+    from modin.pandas import Series
 
     # A pandas label is a hashable, and we exclude the callable, Series and Grouper, which are
     # by values that should not be handled as pandas label of the dataframe.
