+++ src/snowflake/snowpark/modin/plugin/docstrings/series.py
@@ -3277,14 +3277,6 @@ class Series(BasePandasDataset):
         3    2
         dtype: int64
 
-
-        For `Series` axis parameter is unused and defaults to 0.
-
-        >>> ser.take([0, 3], axis=1)
-        0   -1
-        3    2
-        dtype: int64
-
         We may take elements using negative integers for positive indices,
         starting from the end of the object, just like with Python lists.
