+++ src/snowflake/snowpark/modin/plugin/extensions/dataframe_overrides.py
@@ -7,7 +7,7 @@ File containing DataFrame APIs defined in the Modin API layer, but with differen
 pandas, such as `DataFrame.memory_usage`.
 """
 
-from typing import Any
+from typing import Any, Union
 
 import pandas as native_pd
 from modin.pandas import DataFrame
@@ -23,7 +23,7 @@ from snowflake.snowpark.modin.plugin._internal.telemetry import (
 )
 from snowflake.snowpark.modin.plugin.utils.error_message import ErrorMessage
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
-from snowflake.snowpark.modin.utils import _inherit_docstrings
+from snowflake.snowpark.modin.utils import _inherit_docstrings, validate_int_kwarg
 
 
 @_inherit_docstrings(native_pd.DataFrame.memory_usage, apilink="pandas.DataFrame")
@@ -111,6 +111,29 @@ def plot(
         "DataFrame.plot materializes data to the local machine for plotting."
     )
     return self._to_pandas().plot
+
+
+# Upstream modin defines sum differently for series/DF, but we use the same implementation for both.
+@register_dataframe_accessor("sum")
+@snowpark_pandas_telemetry_method_decorator
+def sum(
+    self,
+    axis: Union[Axis, None] = None,
+    skipna: bool = True,
+    numeric_only: bool = False,
+    min_count: int = 0,
+    **kwargs: Any,
+):
+    # TODO: SNOW-1119855: Modin upgrade - modin.pandas.base.BasePandasDataset
+    min_count = validate_int_kwarg(min_count, "min_count")
+    kwargs.update({"min_count": min_count})
+    return self._agg_helper(
+        func="sum",
+        axis=axis,
+        skipna=skipna,
+        numeric_only=numeric_only,
+        **kwargs,
+    )
 
 
 @register_dataframe_accessor("transform")
