+++ src/snowflake/snowpark/modin/plugin/utils/numpy_to_pandas.py
@@ -3,9 +3,9 @@
 #
 from typing import Any, Optional, Union
 
+import modin.pandas as pd
 from modin.pandas.base import BasePandasDataset
 
-import snowflake.snowpark.modin.pandas as pd
 from snowflake.snowpark.modin.pandas.utils import is_scalar
 from snowflake.snowpark.modin.plugin.utils.warning_message import WarningMessage
