+++ tests/integ/modin/series/test_take.py
@@ -3,9 +3,10 @@
 #
 
 import modin.pandas as pd
+import pytest
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
-from tests.integ.modin.sql_counter import SqlCounter
+from tests.integ.modin.sql_counter import SqlCounter, sql_count_checker
 from tests.integ.modin.utils import assert_series_equal
 
 
@@ -33,3 +34,10 @@ def test_series_take():
     expected = pd.Series([6], index=[2])
     with SqlCounter(query_count=2, join_count=2):
         assert_series_equal(actual, expected)
+
+
+@sql_count_checker(query_count=0)
+def test_series_take_axis_1():
+    ser = pd.Series([-1, 5, 6, 2, 4])
+    with pytest.raises(ValueError, match="No axis named 1 for object type Series"):
+        ser.take([0, 1, 3], axis=1)
