+++ tests/integ/modin/test_to_numpy.py
@@ -58,6 +58,7 @@ def test_to_numpy_basic(data, pandas_obj, func):
             assert_array_equal(df.values, native_df.values)
     if pandas_obj == "Series":
         with SqlCounter(query_count=1):
+            # SNOW-1652384 use to_numpy here and test to_list separately
             res = df.to_list()
         expected_res = native_df.to_list()
         for r1, r2 in zip(res, expected_res):
